feat: add in the ability to generate manifests and return them.
So the general idea here would be that there are two ways for a user to
use this.

Firstly and most commonly in CI they will just use:

```
./mill --import ivy:io.kipp::mill-github-dependency-graph0.10:0.0.1-SNAPSHOT show io.kipp.mill.github.dependency.graph.Graph/submit
```
Which will do the full submission. There is also:
```
./mill --import ivy:io.kipp::mill-github-dependency-graph0.10:0.0.1-SNAPSHOT show io.kipp.mill.github.dependency.graph.Graph/generate
```
which a user can use to see what manifests the tool will create to be
submitted.