refactor: clean up how we mark direct at the root level
This just improves the way we mark a direct dependency if it's at the
top level and has already been seen before. It ensure we don't update a
key with the same exact node again hopefully saving us a bit. I also add
in a test here under directRelationship that explicitly tests for this.