fix: ensure ranges don't end up in manifest
There was (and still is) a bit of confusion on the difference between
`retainedVersion` and `reconciledVersion` in couriser. It used to be
that I was using the `retainedVersion`, but that still ends up being a
range, so I switch to `reconciledVersion`. This change also simplifies a
bit of the logic by ensuring the the dependencies listed under
`resolved` are indeed the dependencies that end up being resolved in the
context of your project, not just that single app. I'll explain this
further in the docs.