fix: ensure invalid PURLs are not included
This adds in the package-url-java dep to help ensure that we don't have
invalid written PURLs, however that will still allow for a range, which
we don't want. So we add in some logic that if a range is detected in
the `reconciledVersion`, we throw it out and issue an error. This should
be extremely rare, but this will at least allow the graph to submit and
show.

refs: #40