fix: guard against cyclical dependencies which can cause a stackoverflow
So I don't fully love this solution as it adds to the even more hacks I
need for coursier oddness, but there are situations using deps like you
see in the test here that when you look at the children of the
dependency, it lists itself. Some of them are expected due to different
classifiers, but even one with a classifier will list itself as well as
a dep. This leads to really odd cyclical dependency behavior when you
try to traverse the tree. So to guard against this we do a few
different things:

- we do a better job of not processing children that were already
  processed before
- we filter self out when self is listed as a child
- we also cache lookup to ensure we aren't always going from tree->name
  multiple times as we're iterating.

This will fix
https://github.com/ckipp01/mill-github-dependency-graph/issues/77, but I
think this should probably be addressed upstream. I've created
https://github.com/coursier/coursier/issues/2683 to track this.

Just a NOTE, this still doesn't fix the issues with not having
classifiers shown, but I'll do a follow-up to add those in since it's
not a huge change.

Fixes: https://github.com/ckipp01/mill-github-dependency-graph/issues/77